update: global quest interface and mapping function
