+++ src/lib/global/quest.d.ts
@@ -1,0 +1,12 @@
+interface QuestType {
+    id: number;
+    title: string;
+    taskDescription: string;
+    created: Date;
+    tags: [{
+        id: string;
+        name: string;
+        description: string;
+        }
+    ];
+}
